+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## (IN PROGRESS)
 
+* Search and filter organization exports logs. Refs UIEXPMGR-26.
+
 ## [2.0.0](https://github.com/folio-org/ui-export-manager/tree/v2.0.0) (2021-10-07)
 [Full Changelog](https://github.com/folio-org/ui-export-manager/compare/v1.1.1...v2.0.0)